+++ c-implementation/tab.c
@@ -1,0 +1,194 @@
+// tab.c
+// TAB serial communication protocol header file
+//
+// Written by Bradley Denby
+// Other contributors: None
+//
+// See the top-level LICENSE file for the license.
+
+// Standard library
+#include <stddef.h> // size_t
+#include <stdint.h> // uint8_t, uint32_t
+
+// TAB
+#include <tab.h>    // Header file
+
+// External handler functions
+extern int handle_common_data(common_data_t common_data_buff_i);
+
+// Helper functions
+
+//// Clears rx_cmd_buff data and resets state and indices
+void clear_rx_cmd_buff(rx_cmd_buff_t* rx_cmd_buff_o) {
+  rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_START_BYTE_0;
+  rx_cmd_buff_o->start_index = 0;
+  rx_cmd_buff_o->end_index = 0;
+  for(size_t i=0; i<rx_cmd_buff_o->size; i++) {
+    rx_cmd_buff_o->data[i] = ((uint8_t)0x00);
+  }
+}
+
+//// Clears tx_cmd_buff data and resets state and indices
+void clear_tx_cmd_buff(tx_cmd_buff_t* tx_cmd_buff_o) {
+  tx_cmd_buff_o->empty = 1;
+  tx_cmd_buff_o->start_index = 0;
+  tx_cmd_buff_o->end_index = 0;
+  for(size_t i=0; i<tx_cmd_buff_o->size; i++) {
+    tx_cmd_buff_o->data[i] = ((uint8_t)0x00);
+  }
+}
+
+// Protocol functions
+
+//// Attempts to push byte to end of rx_cmd_buff
+void push_rx_cmd_buff(rx_cmd_buff_t* rx_cmd_buff_o, uint8_t b) {
+  switch(rx_cmd_buff_o->state) {
+    case RX_CMD_BUFF_STATE_START_BYTE_0:
+      if(b==START_BYTE_0) {
+        rx_cmd_buff_o->data[START_BYTE_0_INDEX] = b;
+        rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_START_BYTE_1;
+      }
+      break;
+    case RX_CMD_BUFF_STATE_START_BYTE_1:
+      if(b==START_BYTE_1) {
+        rx_cmd_buff_o->data[START_BYTE_1_INDEX] = b;
+        rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_MSG_LEN;
+      } else {
+        clear_rx_cmd_buff(rx_cmd_buff_o);
+      }
+      break;
+    case RX_CMD_BUFF_STATE_MSG_LEN:
+      if(((uint8_t)0x06)<=b) {
+        rx_cmd_buff_o->data[MSG_LEN_INDEX] = b;
+        rx_cmd_buff_o->start_index = ((uint8_t)0x09);
+        rx_cmd_buff_o->end_index = (b+((uint8_t)0x03));
+        rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_HWID_LSB;
+      } else {
+        clear_rx_cmd_buff(rx_cmd_buff_o);
+      }
+      break;
+    case RX_CMD_BUFF_STATE_HWID_LSB:
+      rx_cmd_buff_o->data[HWID_LSB_INDEX] = b;
+      rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_HWID_MSB;
+      break;
+    case RX_CMD_BUFF_STATE_HWID_MSB:
+      rx_cmd_buff_o->data[HWID_MSB_INDEX] = b;
+      rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_MSG_ID_LSB;
+      break;
+    case RX_CMD_BUFF_STATE_MSG_ID_LSB:
+      rx_cmd_buff_o->data[MSG_ID_LSB_INDEX] = b;
+      rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_MSG_ID_MSB;
+      break;
+    case RX_CMD_BUFF_STATE_MSG_ID_MSB:
+      rx_cmd_buff_o->data[MSG_ID_MSB_INDEX] = b;
+      rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_ROUTE;
+      break;
+    case RX_CMD_BUFF_STATE_ROUTE:
+      rx_cmd_buff_o->data[ROUTE_INDEX] = b;
+      rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_OPCODE;
+      break;
+    case RX_CMD_BUFF_STATE_OPCODE: // no check for valid opcodes (too much)
+      rx_cmd_buff_o->data[OPCODE_INDEX] = b;
+      if(rx_cmd_buff_o->start_index<rx_cmd_buff_o->end_index) {
+        rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_PLD;
+      } else {
+        rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_COMPLETE;
+      }
+      break;
+    case RX_CMD_BUFF_STATE_PLD:
+      if(rx_cmd_buff_o->start_index<rx_cmd_buff_o->end_index) {
+        rx_cmd_buff_o->data[rx_cmd_buff_o->start_index] = b;
+        rx_cmd_buff_o->start_index += 1;
+      }
+      // Must move to COMPLETE state immediately if b is the last byte
+      if(rx_cmd_buff_o->start_index==rx_cmd_buff_o->end_index) {
+        rx_cmd_buff_o->state = RX_CMD_BUFF_STATE_COMPLETE;
+      }
+      break;
+    case RX_CMD_BUFF_STATE_COMPLETE:
+      // If rx_cmd_buff_t holds a complete command, do nothing with new byte b
+      break;
+    default:
+      // Do nothing by default
+      break;
+  }
+}
+
+//// Attempts to clear rx_cmd_buff and populate tx_cmd_buff with reply
+void write_reply(rx_cmd_buff_t* rx_cmd_buff_o, tx_cmd_buff_t* tx_cmd_buff_o) {
+  if(
+   rx_cmd_buff_o->state==RX_CMD_BUFF_STATE_COMPLETE &&
+   tx_cmd_buff_o->empty
+  ) {
+    tx_cmd_buff_o->data[START_BYTE_0_INDEX] = START_BYTE_0;
+    tx_cmd_buff_o->data[START_BYTE_1_INDEX] = START_BYTE_1;
+    tx_cmd_buff_o->data[HWID_LSB_INDEX] = rx_cmd_buff_o->data[HWID_LSB_INDEX];
+    tx_cmd_buff_o->data[HWID_MSB_INDEX] = rx_cmd_buff_o->data[HWID_MSB_INDEX];
+    tx_cmd_buff_o->data[MSG_ID_LSB_INDEX] =
+     rx_cmd_buff_o->data[MSG_ID_LSB_INDEX];
+    tx_cmd_buff_o->data[MSG_ID_MSB_INDEX] =
+     rx_cmd_buff_o->data[MSG_ID_MSB_INDEX];
+    tx_cmd_buff_o->data[ROUTE_INDEX] =
+     (0x0f & rx_cmd_buff_o->data[ROUTE_INDEX]) << 4 |
+     (0xf0 & rx_cmd_buff_o->data[ROUTE_INDEX]) >> 4;
+    // useful variables
+    size_t i = 0;
+    common_data_t common_data_buff = {.size=PLD_MAX_LEN};
+    int success = 0;
+    switch(rx_cmd_buff_o->data[OPCODE_INDEX]) {
+      case COMMON_ACK_OPCODE:
+        tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+        tx_cmd_buff_o->data[OPCODE_INDEX] = COMMON_ACK_OPCODE;
+        break;
+      case COMMON_NACK_OPCODE:
+        tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+        tx_cmd_buff_o->data[OPCODE_INDEX] = COMMON_NACK_OPCODE;
+        break;
+      case COMMON_DEBUG_OPCODE:
+        tx_cmd_buff_o->data[MSG_LEN_INDEX] = rx_cmd_buff_o->data[MSG_LEN_INDEX];
+        tx_cmd_buff_o->data[OPCODE_INDEX] = COMMON_DEBUG_OPCODE;
+        for(i=PLD_START_INDEX; i<rx_cmd_buff_o->end_index; i++) {
+          tx_cmd_buff_o->data[i] = rx_cmd_buff_o->data[i];
+        }
+        break;
+      case COMMON_DATA_OPCODE:
+        // handle common_data
+        for(i=PLD_START_INDEX; i<rx_cmd_buff_o->end_index; i++) {
+          common_data_buff.data[i-PLD_START_INDEX] = rx_cmd_buff_o->data[i];
+        }
+        common_data_buff.end_index = rx_cmd_buff_o->end_index-PLD_START_INDEX;
+        success = handle_common_data(common_data_buff);
+        // reply
+        if(success) {
+          tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+          tx_cmd_buff_o->data[OPCODE_INDEX] = COMMON_ACK_OPCODE;
+        } else {
+          tx_cmd_buff_o->data[MSG_LEN_INDEX] = ((uint8_t)0x06);
+          tx_cmd_buff_o->data[OPCODE_INDEX] = COMMON_NACK_OPCODE;
+        }
+        break;
+      default:
+        break;
+    }
+    tx_cmd_buff_o->end_index = // +((uint8_t)0x03) accounts for 1st 3 bytes
+     (tx_cmd_buff_o->data[MSG_LEN_INDEX]+((uint8_t)0x03));
+    tx_cmd_buff_o->empty = 0;
+    clear_rx_cmd_buff(rx_cmd_buff_o);
+  }
+}
+
+//// Attempts to pop byte from beginning of tx_cmd_buff
+uint8_t pop_tx_cmd_buff(tx_cmd_buff_t* tx_cmd_buff_o) {
+  uint8_t b = 0;
+  if(
+   !tx_cmd_buff_o->empty &&
+   tx_cmd_buff_o->start_index<tx_cmd_buff_o->end_index
+  ) {
+    b = tx_cmd_buff_o->data[tx_cmd_buff_o->start_index];
+    tx_cmd_buff_o->start_index += 1;
+  }
+  if(tx_cmd_buff_o->start_index==tx_cmd_buff_o->end_index) {
+    clear_tx_cmd_buff(tx_cmd_buff_o);
+  }
+  return b;
+}